Routing and blog linking
Moved home to its own router module.
Added a blogHome page that links to the test post. Will have to update
the link list dynamically at some point with all the post pages.
Either that or have a single 'post' page with with dynamically pulled
content depending on the blog post id passed in the URL.
That might be the better way and then I only have to have one blog post
page. That is assuming they will all look the same, which I don't see a
reason why not.