Get Posts route working with call to DB
I tried to get this to work using mongoose.createconnection so that the
application could have the possibility of using more than one database
connection, but I had a lot of troubles with that. When I tried to
create a model on my connection it would say [connection].model is not a
function. I want to figure that out eventually.